@@ -51,10 +51,8 @@ dev = [
     "pytest-mock>=3.12.0",
     
     # Code quality
-    "ruff>=0.2.0",
-    "mypy>=1.8.0",
-    "black>=24.1.0",
-    "isort>=5.13.0",
+    "ruff>=0.14.0",
+    "mypy>=1.13.0",
     
     # Pre-commit
     "pre-commit>=3.6.0",
@@ -92,7 +90,7 @@ markers = [
 
 [tool.ruff]
 line-length = 100
-target-version = "py311"
+target-version = "py313"
 
 [tool.ruff.lint]
 select = [
@@ -104,21 +102,14 @@ select = [
     "C4",  # flake8-comprehensions
     "UP",  # pyupgrade
 ]
-ignore = [
-    "E501",  # line too long (handled by black)
-]
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
 
 [tool.mypy]
-python_version = "3.11"
+python_version = "3.13"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
 ignore_missing_imports = true
-
-[tool.black]
-line-length = 100
-target-version = ["py311"]
-
-[tool.isort]
-profile = "black"
-line_length = 100
