@@ -2,7 +2,7 @@
 
 import logging
 import time
-from typing import Callable
+from collections.abc import Callable
 
 from fastapi import Request, Response, status
 from fastapi.responses import JSONResponse
@@ -24,7 +24,7 @@ async def logging_middleware(request: Request, call_next: Callable) -> Response:
 
     # Log request
     logger.info(
-        f"Request started",
+        "Request started",
         extra={
             "method": request.method,
             "path": request.url.path,
@@ -40,7 +40,7 @@ async def logging_middleware(request: Request, call_next: Callable) -> Response:
 
     # Log response
     logger.info(
-        f"Request completed",
+        "Request completed",
         extra={
             "method": request.method,
             "path": request.url.path,
