@@ -8,8 +8,7 @@ class SummarizeRiskRequest(BaseModel):
 
     ticker: str = Field(..., description="Stock ticker symbol (e.g., PLD, AMT)")
     force_refresh: bool = Field(
-        default=False,
-        description="Force refresh even if cached result exists"
+        default=False, description="Force refresh even if cached result exists"
     )
 
     @field_validator("ticker")
